extract cert configs to a separate package
