+++ hooks/go/utils/certs.go
@@ -1,4 +1,4 @@
-package certs
+package utils
 
 import (
 	"errors"
