+++ openapi/config-values.yaml
@@ -66,4 +66,3 @@ properties:
         type: string
         default: "0 3 * * *"
         description: Backup schedule
-        description: Backup schedule
